Upgrade config to com.styra.opa:springboot:1.0.0
