+++ opa-springboot-demo/after/build.gradle
@@ -1,6 +1,6 @@
 plugins {
 	id 'java'
-	id 'org.springframework.boot' version '3.3.2'
+	id 'org.springframework.boot' version '3.3.5'
 	id 'io.spring.dependency-management' version '1.1.6'
 }
 
@@ -21,11 +21,9 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-web'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-	implementation 'com.fasterxml.jackson.core:jackson-databind:2.17.0'
 	implementation 'org.springframework.security:spring-security-config'
 	implementation 'org.springframework.security:spring-security-web'
-	implementation group: 'com.styra.opa', name: 'springboot', version: '0.0.4'
-	implementation group: 'com.styra', name: 'opa', version: '+'
+	implementation group: 'com.styra.opa', name: 'springboot', version: '1.0.0'
 }
 
 tasks.named('test') {
