+++ opa-springboot-demo/after/src/main/resources/application.properties
@@ -1,1 +1,4 @@
 spring.application.name=demo
+
+opa.url=http://localhost:8181
+opa.path=demo/main
